# BionicInventory.API/Swagger.json
{
...
  "swagger": "2.0",
  "info": {
    "title": "My Title",
    "version": "1.0.0"
  },
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/api/shipments/salesorders/{customerOrderId}": {
      "get": {
        "tags": [
          "Shipments"
        ],
        "operationId": "Shipments_GetCustomerOrderShipmentSummary",
        "parameters": [
          {
            "type": "integer",
            "name": "customerOrderId",
            "in": "path",
            "required": true,
            "x-nullable": false
          },
          {
            "type": "string",
            "name": "type",
            "in": "query",
            "default": "DETAIL",
            "x-nullable": true
          }
        ],
        "responses": {
          "200": {
            "x-nullable": true,
            "description": "",
            "schema": {
              "type": "file"
            }
          }
        }
      },
      "post": {
        "tags": [
          "Shipments"
        ],
        "operationId": "Shipments_CreateNewCustomerOrderShipment",
        "parameters": [
          {
            "type": "integer",
            "name": "customerOrderId",
            "in": "path",
            "required": true,
            "x-nullable": false
          },
          {
            "name": "newShipment",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/NewShipmentDto"
            },
            "x-nullable": true
          }
        ],
        "responses": {
          "200": {
            "x-nullable": true,
            "description": "",
            "schema": {
              "type": "file"
            }
          }
        }
      }
    },
...
    "/api/salesorders/{id}": {
      "get": {
        "tags": [
          "SalesOrder"
        ],
        "operationId": "SalesOrder_GetSalesOrderById",
        "parameters": [
          {
            "type": "integer",
            "name": "id",
            "in": "path",
            "required": true,
            "x-nullable": false
          },
          {
            "type": "string",
            "name": "type",
            "in": "query",
            "default": "VIEW",
            "x-nullable": true
          }
        ],
        "responses": {
          "200": {
            "x-nullable": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/CustomerOrder"
            }
          },
          "400": {
            "description": ""
          },
          "404": {
            "description": ""
          },
          "500": {
            "description": ""
          }
        }
      },
      "put": {
        "tags": [
          "SalesOrder"
        ],
        "operationId": "SalesOrder_UpdateCustomerOrderStatus",
        "parameters": [
          {
            "type": "integer",
            "name": "id",
            "in": "path",
            "required": true,
            "x-nullable": false
          },
          {
            "name": "newStatus",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/StatusUpdateDto"
            },
            "x-nullable": true
          }
        ],
        "responses": {
          "204": {
            "description": ""
          },
          "400": {
            "description": ""
          },
          "404": {
            "description": ""
          },
          "422": {
            "description": ""
          },
          "500": {
            "description": ""
          }
        }
      },
      "delete": {
        "tags": [
          "SalesOrder"
        ],
        "operationId": "SalesOrder_DeleteCustomerOrder",
        "parameters": [
          {
            "type": "integer",
            "name": "id",
            "in": "path",
            "required": true,
            "x-nullable": false
          }
        ],
        "responses": {
          "204": {
            "description": ""
          },
          "404": {
            "description": ""
          },
          "500": {
            "description": ""
          }
        }
      }
    },
    "/api/salesorders/reports": {
      "get": {
        "tags": [
          "SalesOrder"
        ],
        "operationId": "SalesOrder_GetSalesReport",
        "parameters": [
          {
            "type": "string",
            "name": "type",
            "in": "query",
            "default": "YEARLY",
            "x-nullable": true
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "400": {
            "description": ""
          },
          "500": {
            "description": ""
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "Invoice": {
      "type": "object",
      "additionalProperties": false,
      "required": [
        "Id",
        "PurchaseOrderId",
        "PreparedBy",
        "Tax",
        "Discount"
      ],
      "properties": {
        "Id": {
          "type": "integer"
        },
        "PurchaseOrderId": {
          "type": "integer"
        },
        "PreparedBy": {
          "type": "integer"
        },
        "PrintCount": {
          "type": "integer",
          "format": "byte"
        },
        "InvoiceType": {
          "type": "string"
        },
        "PaymentMethod": {
          "type": "string"
        },
        "Tax": {
          "type": "number",
          "format": "double"
        },
        "Note": {
          "type": "string"
        },
        "Discount": {
          "type": "number",
          "format": "double"
        },
        "DateAdded": {
          "type": "string",
          "format": "date-time"
        },
        "DateUpdated": {
          "type": "string",
          "format": "date-time"
        },
        "DueDate": {
          "type": "string",
          "format": "date-time"
        },
        "PreparedByNavigation": {
          "$ref": "#/definitions/Employee"
        },
        "CustomerOrder": {
          "$ref": "#/definitions/CustomerOrder"
        },
        "InvoiceDetail": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/InvoiceDetail"
          }
        },
        "InvoicePayments": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/InvoicePayments"
          }
        }
      }
    },
    "Employee": {
      "type": "object",
      "additionalProperties": false,
      "required": [
        "Id"
      ],
      "properties": {
        "Id": {
          "type": "integer"
        },
        "FirstName": {
          "type": "string"
        },
        "LastName": {
          "type": "string"
        },
        "DateAdded": {
          "type": "string",
          "format": "date-time"
        },
        "DateUpdated": {
          "type": "string",
          "format": "date-time"
        },
        "BookedStockItems": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/BookedStockItems"
          }
        },
        "FinishedProductRecievedByNavigation": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/FinishedProduct"
          }
        },
        "FinishedProductSubmittedByNavigation": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/FinishedProduct"
          }
        },
        "Invoice": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Invoice"
          }
        },
        "InvoicePayments": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/InvoicePayments"
          }
        },
        "ProductionOrderList": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ProductionOrderList"
          }
        },
        "CustomerOrder": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/CustomerOrder"
          }
        },
        "Shipment": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Shipment"
          }
        }
      }
    },
    "BookedStockItems": {
      "type": "object",
      "additionalProperties": false,
      "required": [
        "Id",
        "StockId",
        "BookedBy",
        "BookedFor"
      ],
      "properties": {
        "Id": {
          "type": "integer"
        },
        "StockId": {
          "type": "integer"
        },
        "Canceled": {
          "$ref": "#/definitions/SByte"
        },
        "BookedBy": {
          "type": "integer"
        },
        "DateAdded": {
          "type": "string",
          "format": "date-time"
        },
        "DateUpdated": {
          "type": "string",
          "format": "date-time"
        },
        "BookedFor": {
          "type": "integer"
        },
        "Note": {
          "type": "string"
        },
        "BookedByNavigation": {
          "$ref": "#/definitions/Employee"
        },
        "BookedForNavigation": {
          "$ref": "#/definitions/CustomerOrderItem"
        },
        "Stock": {
          "$ref": "#/definitions/FinishedProduct"
        }
      }
    },
    "SByte": {
      "type": "object",
      "additionalProperties": false
    },
    "CustomerOrderItem": {
      "type": "object",
      "additionalProperties": false,
      "required": [
        "Id",
        "CustomerOrderId",
        "ItemId",
        "Quantity",
        "PricePerItem",
        "DueDate"
      ],
      "properties": {
        "Id": {
          "type": "integer"
        },
        "CustomerOrderId": {
          "type": "integer"
        },
        "ItemId": {
          "type": "integer"
        },
        "Quantity": {
          "type": "integer"
        },
        "PricePerItem": {
          "type": "number",
          "format": "double"
        },
        "DateAdded": {
          "type": "string",
          "format": "date-time"
        },
        "DateUpdated": {
          "type": "string",
          "format": "date-time"
        },
        "DueDate": {
          "type": "string",
          "format": "date-time"
        },
        "CustomerOrder": {
          "$ref": "#/definitions/CustomerOrder"
        },
        "Item": {
          "$ref": "#/definitions/Item"
        },
        "ProductionOrderList": {
          "$ref": "#/definitions/ProductionOrderList"
        },
        "BookedStockBatch": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/BookedStockBatch"
          }
        },
        "BookedStockItems": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/BookedStockItems"
          }
        },
        "InvoiceDetail": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/InvoiceDetail"
          }
        },
        "ShipmentDetail": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ShipmentDetail"
          }
        }
      }
    },
    "CustomerOrder": {
      "type": "object",
      "additionalProperties": false,
      "required": [
        "Id",
        "ClientId",
        "CreatedBy"
      ],
      "properties": {
        "Id": {
          "type": "integer"
        },
        "ClientId": {
          "type": "integer"
        },
        "DateAdded": {
          "type": "string",
          "format": "date-time"
        },
        "DateUpdated": {
          "type": "string",
          "format": "date-time"
        },
        "CreatedBy": {
          "type": "integer"
        },
        "Description": {
          "type": "string"
        },
        "OrderStatus": {
          "type": "string"
        },
        "DueDate": {
          "type": "string",
          "format": "date-time"
        },
        "Client": {
          "$ref": "#/definitions/Customer"
        },
        "CreatedByNavigation": {
          "$ref": "#/definitions/Employee"
        },
        "CustomerOrderItem": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/CustomerOrderItem"
          }
        },
        "Invoice": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Invoice"
          }
        },
        "Shipment": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Shipment"
          }
        }
      }
    },
    "Customer": {
      "type": "object",
      "additionalProperties": false,
      "required": [
        "Id"
      ],
      "properties": {
        "Id": {
          "type": "integer"
        },
        "FullName": {
          "type": "string"
        },
        "Tin": {
          "type": "string"
        },
        "Email": {
          "type": "string"
        },
        "Type": {
          "type": "string"
        },
        "DateAdded": {
          "type": "string",
          "format": "date-time"
        },
        "DateUpdated": {
          "type": "string",
          "format": "date-time"
        },
        "CreditLimit": {
          "type": "number",
          "format": "double"
        },
        "PaymentPeriod": {
          "type": "integer",
          "format": "int32"
        },
        "Fax": {
          "type": "string"
        },
        "PoBox": {
          "type": "string"
        },
        "SocialMedia": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/SocialMedia"
          }
        },
        "Address": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Address"
          }
        },
        "PhoneNumber": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/PhoneNumber"
          }
        },
        "CustomerOrder": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/CustomerOrder"
          }
        }
      }
    },
    "SocialMedia": {
      "type": "object",
      "additionalProperties": false,
      "required": [
        "Id",
        "ClientId"
      ],
      "properties": {
        "Id": {
          "type": "integer"
        },
        "ClientId": {
          "type": "integer"
        },
        "Address": {
          "type": "string"
        },
        "Site": {
          "type": "string"
        },
        "DateAdded": {
          "type": "string",
          "format": "date-time"
        },
        "DateUpdated": {
          "type": "string",
          "format": "date-time"
        },
        "Client": {
          "$ref": "#/definitions/Customer"
        }
      }
    },
    "Address": {
      "type": "object",
      "additionalProperties": false,
      "required": [
        "Id",
        "ClientId"
      ],
      "properties": {
        "Id": {
          "type": "integer"
        },
        "ClientId": {
          "type": "integer"
        },
        "SubCity": {
          "type": "string"
        },
        "City": {
          "type": "string"
        },
        "Country": {
          "type": "string"
        },
        "PhoneNumber": {
          "type": "string"
        },
        "Location": {
          "type": "string"
        },
        "Client": {
          "$ref": "#/definitions/Customer"
        }
      }
    },
    "PhoneNumber": {
      "type": "object",
      "additionalProperties": false,
      "required": [
        "Id",
        "ClientId"
      ],
      "properties": {
        "Id": {
          "type": "integer"
        },
        "ClientId": {
          "type": "integer"
        },
        "Number": {
          "type": "string"
        },
        "DateAdded": {
          "type": "string",
          "format": "date-time"
        },
        "DateUpdated": {
          "type": "string",
          "format": "date-time"
        },
        "Type": {
          "type": "string"
        },
        "Client": {
          "$ref": "#/definitions/Customer"
        }
      }
    },
    "Shipment": {
      "type": "object",
      "additionalProperties": false,
      "required": [
        "Id",
        "BookedBy",
        "DeliveryDate",
        "CustomerOrderId"
      ],
      "properties": {
        "Id": {
          "type": "integer"
        },
        "BookedBy": {
          "type": "integer"
        },
        "ShipmentNote": {
          "type": "string"
        },
        "DeliveryDate": {
          "type": "string",
          "format": "date-time"
        },
        "DateUpdated": {
          "type": "string",
          "format": "date-time"
        },
        "DateAdded": {
          "type": "string",
          "format": "date-time"
        },
        "CustomerOrderId": {
          "type": "integer"
        },
        "BookedByNavigation": {
          "$ref": "#/definitions/Employee"
        },
        "CustomerOrder": {
          "$ref": "#/definitions/CustomerOrder"
        },
        "ShipmentDetail": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ShipmentDetail"
          }
        }
      }
    },
...
    "FinishedProduct": {
      "type": "object",
      "additionalProperties": false,
      "required": [
        "Id",
        "OrderId",
        "SubmittedBy",
        "RecievedBy"
      ],
      "properties": {
        "Id": {
          "type": "integer"
        },
        "OrderId": {
          "type": "integer"
        },
        "DateAdded": {
          "type": "string",
          "format": "date-time"
        },
        "DateUpdated": {
          "type": "string",
          "format": "date-time"
        },
        "SubmittedBy": {
          "type": "integer"
        },
        "RecievedBy": {
          "type": "integer"
        },
        "Quantity": {
          "type": "integer",
          "format": "int32"
        },
        "Order": {
          "$ref": "#/definitions/ProductionOrderList"
        },
        "RecievedByNavigation": {
          "$ref": "#/definitions/Employee"
        },
        "SubmittedByNavigation": {
          "$ref": "#/definitions/Employee"
        },
        "BookedStockItems": {
          "$ref": "#/definitions/BookedStockItems"
        },
        "ShipmentDetail": {
          "$ref": "#/definitions/ShipmentDetail"
        }
      }
    },
    "ProductionOrderList": {
      "type": "object",
      "additionalProperties": false,
      "required": [
        "Id",
        "ItemId",
        "OrderedBy",
        "Quantity",
        "CostPerItem",
        "Start",
        "DueDate"
      ],
      "properties": {
        "Id": {
          "type": "integer"
        },
        "ItemId": {
          "type": "integer"
        },
        "OrderedBy": {
          "type": "integer"
        },
        "Quantity": {
          "type": "integer"
        },
        "CostPerItem": {
          "type": "number",
          "format": "double"
        },
        "Description": {
          "type": "string"
        },
        "DateAdded": {
          "type": "string",
          "format": "date-time"
        },
        "DateUpdated": {
          "type": "string",
          "format": "date-time"
        },
        "Start": {
          "type": "string",
          "format": "date-time"
        },
        "DueDate": {
          "type": "string",
          "format": "date-time"
        },
        "CustomerOrderItemId": {
          "type": "integer"
        },
        "Item": {
          "$ref": "#/definitions/Item"
        },
        "OrderedByNavigation": {
          "$ref": "#/definitions/Employee"
        },
        "CustomerOrderItem": {
          "$ref": "#/definitions/CustomerOrderItem"
        },
        "FinishedProduct": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/FinishedProduct"
          }
        },
        "BookedStockBatch": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/BookedStockBatch"
          }
        },
        "StockBatch": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/StockBatch"
          }
        }
      }
    },
    "Item": {
      "type": "object",
      "additionalProperties": false,
      "required": [
        "Id",
        "UnitCost",
        "GroupId",
        "PrimaryUomId",
        "DefaultStorageId"
      ],
      "properties": {
        "Id": {
          "type": "integer"
        },
        "Code": {
          "type": "string"
        },
        "Name": {
          "type": "string"
        },
        "Weight": {
          "type": "number",
          "format": "double"
        },
        "UnitCost": {
          "type": "number",
          "format": "double"
        },
        "Photo": {
          "type": "string"
        },
        "GroupId": {
          "type": "integer"
        },
        "IsProcured": {
          "$ref": "#/definitions/SByte"
        },
        "IsInventory": {
          "$ref": "#/definitions/SByte"
        },
        "Price": {
          "type": "number",
          "format": "double"
        },
        "ShelfLife": {
          "type": "integer"
        },
        "MinimumQuantity": {
          "type": "number",
          "format": "double"
        },
        "PrimaryUomId": {
          "type": "integer"
        },
        "DateAdded": {
          "type": "string",
          "format": "date-time"
        },
        "DateUpdate": {
          "type": "string",
          "format": "date-time"
        },
        "DefaultStorageId": {
          "type": "integer"
        },
        "StorageLocation": {
          "$ref": "#/definitions/StorageLocation"
        },
        "Group": {
          "$ref": "#/definitions/ProductGroup"
        },
        "PrimaryUom": {
          "$ref": "#/definitions/UnitOfMeasurment"
        },
        "Bom": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Bom"
          }
        },
        "BomItems": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/BomItems"
          }
        },
        "PurchaseOrderItem": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/PurchaseOrderItem"
          }
        },
        "ProductionOrderList": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ProductionOrderList"
          }
        },
        "CustomerOrderItem": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/CustomerOrderItem"
          }
        },
        "Routing": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Routing"
          }
        },
        "VendorPurchaseTerm": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/VendorPurchaseTerm"
          }
        },
        "StockBatch": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/StockBatch"
          }
        },
        "WriteOff": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/WriteOff"
          }
        }
      }
    },
...
    "BookedStockBatch": {
      "type": "object",
      "additionalProperties": false,
      "required": [
        "Id",
        "Quantity"
      ],
      "properties": {
        "Id": {
          "type": "integer"
        },
        "BatchStorageId": {
          "type": "integer"
        },
        "CustomerOrderId": {
          "type": "integer"
        },
        "ProductionOrderId": {
          "type": "integer"
        },
        "Quantity": {
          "type": "number",
          "format": "double"
        },
        "DateAdded": {
          "type": "string",
          "format": "date-time"
        },
        "DateUpdated": {
          "type": "string",
          "format": "date-time"
        },
        "BatchStorage": {
          "$ref": "#/definitions/StockBatchStorage"
        },
        "CustomerOrder": {
          "$ref": "#/definitions/CustomerOrderItem"
        },
        "ProductionOrder": {
          "$ref": "#/definitions/ProductionOrderList"
        }
      }
    },
...
    "InvoiceDetail": {
      "type": "object",
      "additionalProperties": false,
      "required": [
        "Id",
        "SalesOrderId",
        "InvoiceNo",
        "Quantity",
        "UnitPrice"
      ],
      "properties": {
        "Id": {
          "type": "integer"
        },
        "DateAddded": {
          "type": "string",
          "format": "date-time"
        },
        "DateUpdated": {
          "type": "string",
          "format": "date-time"
        },
        "Discount": {
          "type": "number",
          "format": "double"
        },
        "SalesOrderId": {
          "type": "integer"
        },
        "InvoiceNo": {
          "type": "integer"
        },
        "Quantity": {
          "type": "integer",
          "format": "int32"
        },
        "Note": {
          "type": "string"
        },
        "UnitPrice": {
          "type": "number",
          "format": "double"
        },
        "Tax": {
          "type": "number",
          "format": "double"
        },
        "InvoiceNoNavigation": {
          "$ref": "#/definitions/Invoice"
        },
        "CustomerOrderItem": {
          "$ref": "#/definitions/CustomerOrderItem"
        }
      }
    },
    "InvoicePayments": {
      "type": "object",
      "additionalProperties": false,
      "required": [
        "Id",
        "InvoiceNo",
        "Amount",
        "CashierId"
      ],
      "properties": {
        "Id": {
          "type": "integer"
        },
        "InvoiceNo": {
          "type": "integer"
        },
        "Amount": {
          "type": "number",
          "format": "double"
        },
        "Note": {
          "type": "string"
        },
        "DateAdded": {
          "type": "string",
          "format": "date-time"
        },
        "DateUpdated": {
          "type": "string",
          "format": "date-time"
        },
        "CashierId": {
          "type": "integer"
        },
        "PrintCount": {
          "type": "integer",
          "format": "int32"
        },
        "Cashier": {
          "$ref": "#/definitions/Employee"
        },
        "InvoiceNoNavigation": {
          "$ref": "#/definitions/Invoice"
        }
      }
    },
...
    "NewShipmentDto": {
      "allOf": [
        {
          "$ref": "#/definitions/ShipmentDto"
        },
        {
          "type": "object",
          "additionalProperties": false,
          "required": [
            "ShipmentItems"
          ],
          "properties": {
            "ShipmentItems": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/NewShipmentItemDto"
              }
            }
          }
        }
      ]
    },
    "NewShipmentItemDto": {
      "allOf": [
        {
          "$ref": "#/definitions/ShipmentItemDto"
        },
        {
          "type": "object",
          "additionalProperties": false
        }
      ]
    },
    "ShipmentItemDto": {
      "type": "object",
      "x-abstract": true,
      "additionalProperties": false,
      "required": [
        "CustomerOrderItemId",
        "Quantity"
      ],
      "properties": {
        "CustomerOrderItemId": {
          "type": "integer"
        },
        "Quantity": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
    "ShipmentDto": {
      "type": "object",
      "x-abstract": true,
      "additionalProperties": false,
      "required": [
        "deliveryDate",
        "customerOrderId",
        "bookedBy"
      ],
      "properties": {
        "deliveryDate": {
          "type": "string",
          "format": "date-time",
          "minLength": 1
        },
        "customerOrderId": {
          "type": "integer"
        },
        "bookedBy": {
          "type": "integer"
        },
        "shipmentNote": {
          "type": "string"
        }
      }
    },
...
    "StatusUpdateDto": {
      "type": "object",
      "additionalProperties": false,
      "required": [
        "Status"
      ],
      "properties": {
        "Status": {
          "type": "string",
          "minLength": 1
        }
      }
    },
...
  }
}